requirements.txt
Flask==2.0.3
selenium==3.141.0
PyYAML~=6.0
pony==0.7.16
snowballstemmer==2.2.0
Werkzeug==2.2.2
requests
pytest~=8.1.1
Flask-HTTPAuth==4.4.0
